HACK: random: Changes to make compatible with WireGuard
Jason has already stated he will not work around this on his end so
we have to do it here.

get_random_u32 does not have any callers outside of get_random_int
but since get_random_int is in the header file and static so it can
be called elsewhere via an include, we need to rename get_random_u32.

wait_for_random_bytes is not called anywhere outside of random.c so
we can change it to static.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>